requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
websockets==15.0.1
openai==2.3.0
elevenlabs==2.17.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
pandas==2.2.3
python-dotenv==1.1.0
sentence-transformers==5.1.1
langchain-huggingface==0.3.1
langchain-postgres==0.0.15
langchain-openai==0.3.35
langdetect==1.0.9
SQLAlchemy==2.0.40